+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [9.7.1]
+
+### Fixed
+
+- Fixed a build issue on Android when using player SDK v9.7.0.
+
 ## [9.7.0] - 25-07-01
 
 ### Added